refactor: extract AgentShell and its overlays from the agents route
